scripts/scraper/requirements.txt
playwright>=1.49.0
supabase>=2.8.1
python-dotenv>=1.0.1
httpx>=0.27.2
